test: fix dev build proof checks
