+++ src/cli/update-cli/update-command.test.ts
@@ -6,7 +6,6 @@ import { describe, expect, it, vi } from "vitest";
 import { resolveGatewayInstallEntrypoint } from "../../daemon/gateway-entrypoint.js";
 import type { GatewayService } from "../../daemon/service.js";
 import type { UpdateRunResult } from "../../infra/update-runner.js";
-import { defaultRuntime } from "../../runtime.js";
 import {
   updatePluginsAfterCoreUpdate,
   type PostCorePluginUpdateResult,
@@ -22,7 +21,6 @@ import {
 import {
   resolvePostUpdateServiceStateReadEnv,
   resolveUpdatedGatewayRestartPort,
-  maybeRestartService,
   shouldPrepareUpdatedInstallRestart,
 } from "./update-command-service.js";
 import { testing as updateCommandServiceTesting } from "./update-command-service.test-support.js";
@@ -179,32 +177,6 @@ describe("resolveUpdatedGatewayRestartPort", () => {
         serviceEnv: {},
       }),
     ).toBe(19000);
-  });
-});
-
-describe("maybeRestartService", () => {
-  it("reports service ownership skips to JSON callers", async () => {
-    const errorSpy = vi.spyOn(defaultRuntime, "error").mockImplementation(() => undefined);
-
-    await expect(
-      maybeRestartService({
-        shouldRestart: false,
-        result: {
-          status: "ok",
-          mode: "npm",
-          steps: [],
-          durationMs: 0,
-        },
-        channel: "stable",
-        opts: { json: true },
-        refreshServiceEnv: false,
-        gatewayPort: 18789,
-        serviceMutationSkipMessage: "service management skipped: ownership conflict",
-        timeoutMs: 1_000,
-      }),
-    ).resolves.toBe(true);
-
-    expect(errorSpy).toHaveBeenCalledWith("service management skipped: ownership conflict");
   });
 });
